DAC scan analysis moved to utils.anautilities to be more pythonic
